chore: reorganize contract modules
