+++ src/Contracts/Authentication/LoginRequest.cs
@@ -1,7 +1,7 @@
 namespace Contracts.Authentication;
 
 /// <summary>
-/// Request type to login a registered user.
+/// Represents a request to log in.
 /// </summary>
 /// <param name="Email">The user email.</param>
 /// <param name="Password">The user password.</param>
