+++ src/Contracts/Common/PaymentMethods/PaymentMethod.cs
@@ -1,6 +1,6 @@
 using System.Text.Json.Serialization;
 
-namespace Contracts.Orders.Common;
+namespace Contracts.Common.PaymentMethods;
 
 /// <summary>
 /// Represents a polymorphic payment method.
