+++ src/Contracts/Payments/PaymentStatusChangedRequest.cs
@@ -1,8 +1,8 @@
-namespace Contracts.Notifications;
+namespace Contracts.Payments;
 
 /// <summary>
 /// Represents a payment status notification.
 /// </summary>
 /// <param name="PaymentId">The payment id.</param>
 /// <param name="PaymentStatus">The payment status.</param>
-public record PaymentStatusChangedNotification(string PaymentId, string PaymentStatus);
+public record PaymentStatusChangedRequest(string PaymentId, string PaymentStatus);
