+++ src/WebApi/Common/Mappings/PaymentMappingConfig.cs
@@ -1,7 +1,5 @@
 using Application.Payments.Commands.UpdatePaymentStatus;
-
-using Contracts.Notifications;
-
+using Contracts.Payments;
 using Mapster;
 
 namespace WebApi.Common.Mappings;
@@ -14,7 +12,7 @@ public class PaymentMappingConfig : IRegister
     /// <inheritdoc/>
     public void Register(TypeAdapterConfig config)
     {
-        config.NewConfig<PaymentStatusChangedNotification, UpdatePaymentStatusCommand>()
+        config.NewConfig<PaymentStatusChangedRequest, UpdatePaymentStatusCommand>()
             .MapWith(src => new UpdatePaymentStatusCommand(src.PaymentId, src.PaymentStatus));
     }
 }
